backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-anthropic
chromadb
sentence-transformers
pypdf
python-dotenv
pytest
